requirements.txt
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
google-api-core==2.19.2
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
gspread==6.1.2
httplib2==0.22.0
idna==3.8
linkedin_api==2.2.0
lxml==5.3.0
numpy==2.1.0
oauthlib==3.2.2
pandas==2.2.2
proto-plus==1.24.0
protobuf==5.28.0
py4j==0.10.9.7
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.4
pyspark==3.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
soupsieve==2.6
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
